Workaround to handle VCF files that lack both header and variants.
There's a bug in VEP which causes it to create an empty file if no variants are present in the input file. The correct behaviour would be to create a VCF file with a header. This workaround fixes a crash in reportgen when working with a VCF file which is completely empty (no header or variants).